// client/src/types.ts
export interface Employee {
  email: string
  gitHub: string | null
  highlighted: boolean
  imagePortraitUrl: string | null
  imageWallOfLeetUrl: string | null
  linkedIn: string | null
  mainText: string | null
  manager: string | null
  name: string
  office: string | null
  orgUnit: string
  phoneNumber: string | null
  published: boolean
  stackOverflow: string | null
  twitter: string | null
}